update session expires when updating expires key
Catalyst only updates the session key if the data has changed. This allows the expires and session ttl to be different in redis.

If the session data does not change for a long enough period the session key will be expired by redis even tho the user is still using the app.